Big improvements to the code structure and added popularity
Added popularity and all the things needed to set it.
Added a new class for our subjects (not in use YET).

Changed how the teacher is displayed on the agenda.
Changed how the teachers are displayed on the lesson create view,
Changed all the adding to the panes, view, etc. (Saved+-200 lines of code)
Changed some code to match OUR style guide.
Changed some labels in the GUI (mostly typos).

Removed the MainClass since it had no use.
Removed some unused code.
Removed a boat load of typos and errors.